refactor(obra-social): modificacion routes
