--- modules/obraSocial/schemas/periodoPadronesPuco.ts
@@ -1,8 +1,0 @@
-import * as mongoose from 'mongoose';
-
-const periodoPadronesPucoSchema = new mongoose.Schema({
-    id: mongoose.Schema.Types.ObjectId,
-    version: Date
-});
-
-export let periodoPadronesPuco = mongoose.model('periodoPadronesPuco', periodoPadronesPucoSchema, 'periodoPadronesPuco');
